src/requirements.txt
academictorrents==2.3.3
albumentations==0.4.3
basicsr==1.4.2
cog==0.4.3
diffusers==0.2.4
easydict==1.9
einops==0.3.0
ffmpeg==1.4
ftfy==6.1.1
gfpgan==1.3.5
gradio==3.3
horovod==0.25.0
ipdb==0.13.7
jieba==0.42.1
jsonlines==3.1.0
ldm==0.1.3
matplotlib==3.3.4
more_itertools==8.14.0
multilingual_clip==1.0.10
numpy==1.23.3
omegaconf==2.1.1
opencv_python==4.1.2.30
packaging==20.9
pandas==1.1.4
Pillow==9.2.0
pip==21.0.1
pytest==6.2.2
pytorch_lightning==1.7.3
PyYAML==6.0
regex==2021.3.17
requests==2.25.1
scikit_image==0.15.0
setuptools==59.5.0
six==1.15.0
scikit-image
timm==0.6.7
tqdm==4.64.0
transformers==4.29.1
webdataset==0.2.26
taming-transformers-rom1504
clip
realesrgan